refactor(ddgr): move mock handler to test file to fix lint
MockHandler and its state (mockGetErr, mockUpdateErr, etc.) lived in
utils.go but were only referenced from test files, causing golangci-lint
to flag them as unused (and the error vars as violating errname). Moved
all mock code to mock_handler_test.go and registered the handler via
init(), with a testHandlers map in utils.go as the bridge.
